testsuite: Use ovn-macros instead of ofproto-macros.
There are duplicated code in ofproto-macros.at and ovn-macros.at.
It seems ovn-macros.at is supposed to be used for ovn specific
macros, but it is not really used.

This patch removes the duplicated code from ofproto-macros.at and
corrected the path in ovn-macros.at for the ovs/ovn split.

Signed-off-by: Han Zhou <hzhou@ovn.org>