ovn-ctl: Change the order of ovn-lib and ovs-lib.
Some function names are identical in both libs, e.g. daemon_is_running().
Currently we included ovs-lib after ovn-lib, so the one in ovs-lib
overrides the one in ovn-lib, which results in unexpected behavior.

This patch changes the order so that functions in ovn-lib will take
effect when there is confict names.

Signed-off-by: Han Zhou <hzhou@ovn.org>